--- a/src/Adaptive.Aeron.Tests/FragmentAssemblerTest.cs
+++ b/src/Adaptive.Aeron.Tests/FragmentAssemblerTest.cs
@@ -13,7 +13,7 @@ namespace Adaptive.Aeron.Tests
         private const int INITIAL_TERM_ID = 3;
 
         private FragmentHandler delegateFragmentHandler;
-        private IDirectBuffer termBuffer;
+        private UnsafeBuffer termBuffer;
         private Header header;
         private FragmentAssembler adapter;
 
@@ -21,7 +21,7 @@ namespace Adaptive.Aeron.Tests
         public void SetUp()
         {
             delegateFragmentHandler = A.Fake<FragmentHandler>();
-            termBuffer = A.Fake<IDirectBuffer>();
+            termBuffer = A.Fake<UnsafeBuffer>();
             adapter = new FragmentAssembler(delegateFragmentHandler);
             header = A.Fake<Header>(x => x.Wrapping(new Header(INITIAL_TERM_ID, LogBufferDescriptor.TERM_MIN_LENGTH)));
 
@@ -157,7 +157,7 @@ namespace Adaptive.Aeron.Tests
 
             adapter.OnFragment(srcBuffer, offset, length, header);
 
-            A.CallTo(() => delegateFragmentHandler(A<IDirectBuffer>._, A<int>._, A<int>._, A<Header>._)).MustNotHaveHappened();
+            A.CallTo(() => delegateFragmentHandler(A<UnsafeBuffer>._, A<int>._, A<int>._, A<Header>._)).MustNotHaveHappened();
         }
 
         [Test]
@@ -171,7 +171,7 @@ namespace Adaptive.Aeron.Tests
             adapter.OnFragment(srcBuffer, offset, length, header);
             adapter.OnFragment(srcBuffer, offset, length, header);
 
-            A.CallTo(() => delegateFragmentHandler(A<IDirectBuffer>._, A<int>._, A<int>._, A<Header>._)).MustNotHaveHappened();
+            A.CallTo(() => delegateFragmentHandler(A<UnsafeBuffer>._, A<int>._, A<int>._, A<Header>._)).MustNotHaveHappened();
         }
     }
 }