--- a/src/Adaptive.Aeron.Tests/LogBuffer/TermRebuilderTest.cs
+++ b/src/Adaptive.Aeron.Tests/LogBuffer/TermRebuilderTest.cs
@@ -11,12 +11,12 @@ namespace Adaptive.Aeron.Tests.LogBuffer
     {
         private static readonly int TERM_BUFFER_CAPACITY = LogBufferDescriptor.TERM_MIN_LENGTH;
 
-        private IAtomicBuffer _termBuffer;
+        private UnsafeBuffer _termBuffer;
 
         [SetUp]
         public void SetUp()
         {
-            _termBuffer = A.Fake<IAtomicBuffer>();
+            _termBuffer = A.Fake<UnsafeBuffer>();
             A.CallTo(() => _termBuffer.Capacity).Returns(TERM_BUFFER_CAPACITY);
         }