--- a/src/Adaptive.Aeron.sln
+++ b/src/Adaptive.Aeron.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+VisualStudioVersion = 14.0.25123.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Adaptive.Aeron", "Adaptive.Aeron\Adaptive.Aeron.csproj", "{CE06A485-98F5-4FEF-A756-D0EF4DDA8309}"
 EndProject
@@ -40,9 +40,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Adaptive.Aeron.Samples.BufferClaimIpcThroughput", "Samples\Adaptive.Aeron.Samples.BufferClaimIpcThroughput\Adaptive.Aeron.Samples.BufferClaimIpcThroughput.csproj", "{F1EC8358-B42D-4459-B61D-BA26B09A2B86}"
 EndProject
 Global
-	GlobalSection(Performance) = preSolution
-		HasPerformanceSessions = true
-	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64