--- a/src/Adaptive.Aeron.sln
+++ b/src/Adaptive.Aeron.sln
@@ -40,6 +40,9 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Adaptive.Aeron.Samples.BufferClaimIpcThroughput", "Samples\Adaptive.Aeron.Samples.BufferClaimIpcThroughput\Adaptive.Aeron.Samples.BufferClaimIpcThroughput.csproj", "{F1EC8358-B42D-4459-B61D-BA26B09A2B86}"
 EndProject
 Global
+	GlobalSection(Performance) = preSolution
+		HasPerformanceSessions = true
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64