These changes combined push IPC performance to 22 Mops from 14 Mops. It requires some cleaning, so this PR is mostly FYI & WIP. #29 #31