@buybackoff 2016-05-22
With the https://github.com/AdaptiveConsulting/Aeron.NET/pull/37/commits/0f9c0b32fb2e040aef87451718cd338aa1c1caeb commit, the median values for throughput printed lines:
* Non-cached delegate: 22.6 Mops, StDev 0.63, a lot of GC0 for temp delegates
* Cached delegate: 19.9 Mops, StDev 0.18, no garbage at all
* Current master: 13.7 Mops, StDev 0.31, a lot of GC0 for temp delegates

On the same machine, Java version runs at median 23.7 Mops, StDev 0.90.

Main contributions to performance are:
* Interface method calls elimination - other than they are more expensive by themselves, they are not candidates for JIT inlining. This change alone in the main contributor, it boosts throughput to c.20 Mops.
* MemoryCopy was implemented as a DynamicMethod delegate, which is also not a candidate for JIT inlining. Manual memory copy inside `UnsafeBuffer.PutBytes` gives noticeable improvement. (This [post](http://frankniemeyer.blogspot.ru/2014/07/methods-for-reading-structured-binary.html) discusses different memcpy methods, simple "*dst++ = *src++" is already one of the fastest. In the latest commit memcpy was optimized to copy in chunks of 64/32/16/8/4/1 bytes).
* Making UnsafeBuffer sealed and implementing header flyweights as structs, sealing virtual methods in other places.



There is a curios case with a cached delegate that reduces allocations but performance drops with that, I have posted a [question](http://stackoverflow.com/questions/37372719/c-sharp-why-using-instance-method-as-delegate-allocates-gc0-temp-objects-but-10) on SO about this issue.

Ping-pong latency is not affected by delegate caching (that affects only IpcThroughput project):

Pinging 1000000 messages
Histogram of RTT latencies in microseconds.
       Value     Percentile TotalCount 1/(1-Percentile)

      10.591 0.000000000000         50           1.00
      11.871 0.100000000000     301357           1.11
      11.871 0.200000000000     301357           1.25
      11.871 0.300000000000     301357           1.43
      12.191 0.400000000000     437437           1.67
      12.831 0.500000000000     678836           2.00
      12.831 0.550000000000     678836           2.22
      12.831 0.600000000000     678836           2.50
      12.831 0.650000000000     678836           2.86
      13.151 0.700000000000     930516           3.33
      13.151 0.750000000000     930516           4.00
      13.151 0.775000000000     930516           4.44
      13.151 0.800000000000     930516           5.00
      13.151 0.825000000000     930516           5.71
      13.151 0.850000000000     930516           6.67
      13.151 0.875000000000     930516           8.00
      13.151 0.887500000000     930516           8.89
      13.151 0.900000000000     930516          10.00
      13.151 0.912500000000     930516          11.43
      13.151 0.925000000000     930516          13.33
      13.471 0.937500000000     953262          16.00
      13.471 0.943750000000     953262          17.78
      13.471 0.950000000000     953262          20.00
      13.791 0.956250000000     961820          22.86
      14.119 0.962500000000     972590          26.67
      14.119 0.968750000000     972590          32.00
      14.119 0.971875000000     972590          35.56
      14.439 0.975000000000     976122          40.00
      15.079 0.978125000000     978494          45.71
      16.687 0.981250000000     981566          53.33
      20.863 0.984375000000     984523          64.00
      24.703 0.985937500000     985997          71.11
      29.839 0.987500000000     987563          80.00
      35.935 0.989062500000     989113          91.43
      44.287 0.990625000000     990654         106.67
      61.279 0.992187500000     992190         128.00
      75.071 0.992968750000     992978         142.22
      95.935 0.993750000000     993751         160.00
     117.759 0.994531250000     994614         182.86
     120.639 0.995312500000     995320         213.33
     125.759 0.996093750000     996106         256.00
     129.279 0.996484375000     996510         284.44
     134.399 0.996875000000     996896         320.00
     142.207 0.997265625000     997276         365.71
     153.343 0.997656250000     997664         426.67
     169.727 0.998046875000     998049         512.00
     182.527 0.998242187500     998244         568.89
     201.215 0.998437500000     998439         640.00
     221.695 0.998632812500     998634         731.43
     258.943 0.998828125000     998830         853.33
     324.095 0.999023437500     999024        1024.00
     364.287 0.999121093750     999122        1137.78
     396.287 0.999218750000     999220        1280.00
     422.911 0.999316406250     999317        1462.86
     446.207 0.999414062500     999416        1706.67
     468.735 0.999511718750     999512        2048.00
     479.231 0.999560546875     999562        2275.56
     490.239 0.999609375000     999612        2560.00
     503.039 0.999658203125     999659        2925.71
     525.311 0.999707031250     999708        3413.33
     544.767 0.999755859375     999761        4096.00
     552.447 0.999780273438     999782        4551.11
     564.223 0.999804687500     999805        5120.00
     587.775 0.999829101563     999831        5851.43
     606.207 0.999853515625     999854        6826.67
     641.535 0.999877929688     999878        8192.00
     660.479 0.999890136719     999891        9102.22
     694.271 0.999902343750     999903       10240.00
     741.887 0.999914550781     999915       11702.86
     786.943 0.999926757813     999927       13653.33
     889.855 0.999938964844     999939       16384.00
     963.071 0.999945068359     999946       18204.44
    1018.367 0.999951171875     999952       20480.00
    1104.895 0.999957275391     999958       23405.71
    1205.247 0.999963378906     999964       27306.67
    1394.687 0.999969482422     999970       32768.00
    1618.943 0.999972534180     999973       36408.89
    1657.855 0.999975585938     999976       40960.00
    1854.463 0.999978637695     999979       46811.43
    1946.623 0.999981689453     999982       54613.33
    1993.727 0.999984741211     999985       65536.00
    2004.991 0.999986267090     999987       72817.78
    2055.167 0.999987792969     999988       81920.00
    2081.791 0.999989318848     999990       93622.86
    2084.863 0.999990844727     999991      109226.67
    2093.055 0.999992370605     999993      131072.00
    2119.679 0.999993133545     999994      145635.56
    2119.679 0.999993896484     999994      163840.00
    2207.743 0.999994659424     999995      187245.71
    2226.175 0.999995422363     999996      218453.33
    2861.055 0.999996185303     999997      262144.00
    2861.055 0.999996566772     999997      291271.11
    2861.055 0.999996948242     999997      327680.00
    2889.727 0.999997329712     999998      374491.43
    2889.727 0.999997711182     999998      436906.67
    3211.263 0.999998092651     999999      524288.00
    3211.263 0.999998283386     999999      582542.22
    3211.263 0.999998474121     999999      655360.00
    3211.263 0.999998664856     999999      748982.86
    3211.263 0.999998855591     999999      873813.33
    3887.103 0.999999046326    1000000
    3887.103 1.000000000000    1000000
#[Mean    =       14.064, StdDeviation   =       22.107]
#[Max     =     3887.103, Total count    =      1000000]
#[Buckets =           24, SubBuckets     =         2048]