@buybackoff 2016-05-22
It was a fun challenge, but at the same time quite tiresome to do JIT compiler job for it... Obviously .NET JIT is lazy with many things. I will now stop adding any new changes, current version is on par with Java. This is a kind of kitchen-sink commit, so please cherry pick the parts you think are appropriate, or suggest what other changes are needed to merge it?